Not working at all. Trying to redo it so that Map does not have any drawing informatin. It is hard
