+++ src/com/TigersIter2/views/TileView.java
@@ -1,4 +1,4 @@
-package com.TigersIter2.maps;
+package com.TigersIter2.views;
 
 import com.TigersIter2.assets.StaticVar;
 import com.TigersIter2.maps.terrains.TerrainType;
@@ -9,17 +9,17 @@ import java.awt.*;
 /**
  * Created by Miles on 2/28/16.
  */
-public class Tile extends JComponent {
+public class TileView extends JComponent {
 
     Graphics2D g2d;
     int currentXLocation, currentYLocation;
     TerrainType terrainType;
 
-    public Tile(TerrainType t){  //Constructor specifies stuuuuufff.... I'm terrible at commenting code, you guys (Miles)
+    public TileView(TerrainType t){  //Constructor specifies stuuuuufff.... I'm terrible at commenting code, you guys (Miles)
         terrainType = t;
     }
 
-    public Tile(int x, int y, TerrainType t){
+    public TileView(int x, int y, TerrainType t){
         currentXLocation = x;
         currentYLocation = y;
         terrainType = t;
@@ -33,11 +33,16 @@ public class Tile extends JComponent {
          */
 
         //TODO: Make this a lot nicer if possible!!!
-        if((currentXLocation % 2) == 0) { //If X is even, draw it one way.
-            g2d.drawImage(terrainType.getTerrainImage(), Math.round(currentXLocation * StaticVar.terrainImageWidth * .75f) - (StaticVar.terrainImageWidth/2), (currentYLocation * StaticVar.terrainImageHeight) - (StaticVar.terrainImageHeight/2), null);
+
+        if(currentXLocation > 22 || currentYLocation > 10){
+            //do nothing - we're out of range!
         }
-        else{
-            g2d.drawImage(terrainType.getTerrainImage(), Math.round(currentXLocation * StaticVar.terrainImageWidth * .75f) - (StaticVar.terrainImageWidth/2), (currentYLocation * StaticVar.terrainImageHeight), null);
+        else {
+            if ((currentXLocation % 2) == 0) { //If X is even, draw it one way.
+                g2d.drawImage(terrainType.getTerrainImage(), Math.round(currentXLocation * StaticVar.terrainImageWidth * .75f) - (StaticVar.terrainImageWidth / 2), (currentYLocation * StaticVar.terrainImageHeight) - (StaticVar.terrainImageHeight / 2), null);
+            } else {
+                g2d.drawImage(terrainType.getTerrainImage(), Math.round(currentXLocation * StaticVar.terrainImageWidth * .75f) - (StaticVar.terrainImageWidth / 2), (currentYLocation * StaticVar.terrainImageHeight), null);
+            }
         }
     }
